bios/volt: handle voltage table version 0x50 with 0ed header
Some Fermi+ gpus have no usefull header in the voltage table, which means
nouveau has to read the voltages out of the entries directly.

The mask may be bigger than 0x1fffff, but this value is already >2V, so it will
be fine for now.

This patch fixes volting issues on those cards enabling them to switch cstates.

Signed-off-by: Karol Herbst <nouveau@karolherbst.de>
Reviewed-by: Martin Peres <martin.peres@free.fr>
Tested-by: Pierre Moreau <pierre.morrow@free.fr>