therm: debug print the FSRM config on [nvc0,nvf0) cards
before we can let nouveau fully reclock the gpu automatically, we should have
a fail safe mechanism to prevent the card from overheating.

Starting with G84 GPUs the so called FSRM does the job and automatically
downclocks the GPU when specific temperatures thresholds are hit. Lucky for us
the needed bits are already configured through a script inside the vbios at
POST time and we only need to verify the set bits.

This Patch implements the verification steps for all [nvc0,nvf0) cards by
printing a debug message and in the case we detect a wrongly configured fsrm
we also print an error.

Sadly on nvf0 and newer cards this looks a bit different and a other patch for
cards older than nvc0 will follow the next days.

Signed-off-by: Karol Herbst <nouveau@karolherbst.de>