pmu/fuc: read out counters and store them
The counter values are read out periodically and divided by the total
amount of ticks divided by 255 to have nice 8 bit values. A higher
precision won't be needed. This way we can have simple PMU-host
interfaces.

The polling is designed in such a way so that we can change the polling
period later if we need it.

Signed-off-by: Karol Herbst <karolherbst@gmail.com>